Fix code gen on arithmetic shift or bitwise not

+++ tests/xy_c_compile_resources/bitwiseOperations/bitBasics.c
@@ -10,6 +10,6 @@ uint32_t bitBasics_test(void) {
     const uint32_t l_e = l_a ^ l_b;
     const uint32_t l_f = l_d << 4 ^ l_e >> 5;
     const uint32_t l_g = ~(l_c | l_f & l_e);
-    const uint32_t l_h = (int32_t)l_g >> 10;
+    const uint32_t l_h = (uint32_t)((int32_t)l_g >> 10);
     return (uint32_t)l_h;
 }
